Declare specialization constants in the AST
A new qualifier (EvqSpecConst) is added so that specialization constants
can be declared in tree.

This enables TVariable references to specialization constants to be
validated, which were in fact invalid as every reference declared a new
variable.  That is fixed.

Bug: angleproject:4889
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/2941451
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>

+++ src/libANGLE/renderer/vulkan/vk_helpers.cpp
@@ -7777,10 +7777,10 @@ void ShaderProgramHelper::setSpecializationConstant(sh::vk::SpecializationConsta
             mSpecializationConstants.surfaceRotation = value;
             break;
         case sh::vk::SpecializationConstantId::DrawableWidth:
-            mSpecializationConstants.drawableWidth = value;
+            mSpecializationConstants.drawableWidth = static_cast<float>(value);
             break;
         case sh::vk::SpecializationConstantId::DrawableHeight:
-            mSpecializationConstants.drawableHeight = value;
+            mSpecializationConstants.drawableHeight = static_cast<float>(value);
             break;
         default:
             UNREACHABLE();
