+++ src/libANGLE/renderer/vulkan/vk_utils.h
@@ -727,8 +727,8 @@ struct SpecializationConstants final
 {
     VkBool32 lineRasterEmulation;
     uint32_t surfaceRotation;
-    uint32_t drawableWidth;
-    uint32_t drawableHeight;
+    float drawableWidth;
+    float drawableHeight;
 };
 ANGLE_DISABLE_STRUCT_PADDING_WARNINGS
